+++ inst/tinytest/test-escape.R
@@ -14,3 +14,9 @@ expect_snapshot_print(tab, "escape-latex")
 dat <- data.frame("blah_blah" = 1:2)
 tab <- tt(dat) |> format_tt(escape = TRUE) |> save_tt("latex")
 expect_snapshot_print(tab, "escape-latex_colnames")
+
+
+# Issue #150: avoid double escaping captions
+tab <- tt(head(iris), caption = "Blah blah \\label{tab:blah-blah}")
+tab@output <- "latex"
+expect_snapshot_print(tab, "escape-issue150_caption_escape")