+++ NEWS.md
@@ -24,6 +24,7 @@ Misc:
 * New `tinytable_tt_digits` global option can set the default number of digits in the `tt()` function.
 * Refactor: `tinytable` objects are now S4 class objects, with slots to hold data about the content and structure.
 * `as.character()` now works on `tinytable` objects, choosing the output format in the `@output` slot by default.
+* LaTeX code in captions no longer requires double escaping, allowing: `tt(x, caption = "Blah blah \\label{tab:blah})`
 
 Breaking changes: